+++ characterscan/__init__.py
@@ -1,4 +1,4 @@
 """Initialize the app"""
 
-__version__ = "1.16.3"
+__version__ = "1.17.0"
 __title__ = "Character Scan"